Move animation related classes to the "animation" package
