+++ subscribing-example-app/src/main/java/com/ably/tracking/example/subscriber/MainActivity.kt
@@ -14,14 +14,14 @@ import androidx.transition.TransitionManager
 import com.ably.tracking.Accuracy
 import com.ably.tracking.Resolution
 import com.ably.tracking.TrackableState
-import com.ably.tracking.ui.CoreLocationAnimator
-import com.ably.tracking.ui.LocationAnimator
-import com.ably.tracking.ui.Position
 import com.ably.tracking.connection.Authentication
 import com.ably.tracking.connection.ConnectionConfiguration
 import com.ably.tracking.logging.LogHandler
 import com.ably.tracking.logging.LogLevel
 import com.ably.tracking.subscriber.Subscriber
+import com.ably.tracking.ui.animation.CoreLocationAnimator
+import com.ably.tracking.ui.animation.LocationAnimator
+import com.ably.tracking.ui.animation.Position
 import com.google.android.gms.maps.CameraUpdateFactory
 import com.google.android.gms.maps.GoogleMap
 import com.google.android.gms.maps.SupportMapFragment
